// frontend/src/api/openapi/models/UserCreate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserCreate = {
  /**
   * E-Mail of the User
   */
  email: string;
  /**
   * First name of the User
   */
  first_name: string;
  /**
   * Last name of the User
   */
  last_name: string;
  /**
   * Hashed password of the User
   */
  password: string;
};